backend/foodest_backend/requirements.txt
asgiref==3.7.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django
django-cors-headers
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework
djangorestframework-simplejwt
djoser
drf-extra-fields==3.4.0
flake8==5.0.4
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
drf_spectacular
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
gunicorn==20.1.0
